Fixed some bugs noticed when compiling on windows.
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4361 941e2c2b-deac-454f-805a-451daa25f33c

+++ libhfst/src/implementations/HfstTransitionGraph.h
@@ -3653,7 +3653,7 @@
            std::set<HfstState> epsilon_path_states;
            epsilon_path_states.insert(0);
            unsigned int index=0;
-           HfstOneLevelPath path(0, s);
+           HfstOneLevelPath path((float)0, s);
 
            return is_lookup_infinitely_ambiguous(path, index, INITIAL_STATE,
                                                  epsilon_path_states);
